requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==46.0.5
distro==1.9.0
Django==5.2.12
django-celery-beat==2.9.0
django-environ==0.13.0
django-timezone-field==7.2.1
djangorestframework==3.16.1
google-auth==2.49.1
google-auth-oauthlib==1.3.0
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
kombu==5.6.2
oauthlib==3.3.1
openai==2.26.0
packaging==26.0
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyaes==1.6.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-crontab==3.3.0
python-dateutil==2.9.0.post0
redis==7.3.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
Telethon==1.42.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
update-checker==0.18.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
websocket-client==1.9.0
